Remove sqlite picker
- Disable cachepicker (performance)
- Fix battery pick
