+++ src/faebryk/libs/picker/api/pickers.py
@@ -6,7 +6,7 @@ import faebryk.libs.picker.api.picker_lib as picker_lib
 from faebryk.core.module import Module
 from faebryk.core.solver.solver import Solver
 from faebryk.libs.picker.api.api import ApiHTTPError
-from faebryk.libs.picker.common import StaticPartPicker
+from faebryk.libs.picker.api.common import StaticPartPicker
 from faebryk.libs.picker.jlcpcb.jlcpcb import Component
 from faebryk.libs.picker.picker import PickError
 
@@ -30,9 +30,9 @@ class StaticApiPartPicker(StaticPartPicker):
     def _find_parts(self, module: Module, solver: Solver) -> list[Component]:
         match self.mfr, self.mfr_pn, self.lcsc_pn:
             case (mfr, mfr_pn, None) if mfr is not None and mfr_pn is not None:
-                return [picker_lib._find_component_by_mfr(mfr, mfr_pn)]
+                return [picker_lib.find_component_by_mfr(mfr, mfr_pn)]
             case (None, None, lcsc_pn) if lcsc_pn is not None:
-                return [picker_lib._find_component_by_lcsc_id(lcsc_pn)]
+                return [picker_lib.find_component_by_lcsc_id(lcsc_pn)]
             case (None, None, None):
                 raise PickError("No parameters provided", module)
         return []
