+++ src/faebryk/libs/picker/jlcpcb/jlcpcb.py
@@ -2,22 +2,13 @@
 # SPDX-License-Identifier: MIT
 
 import asyncio
-import datetime
 import json
 import logging
-import math
-import os
-import struct
-import sys
 from dataclasses import dataclass, field
 from pathlib import Path
 from textwrap import indent
-from typing import Any, Callable, Generator, Self, Sequence
+from typing import Any, Callable
 
-import patoolib
-import requests
-from rich.progress import track
-from tortoise import Tortoise
 from tortoise.expressions import Q
 from tortoise.fields import CharField, DatetimeField, IntField, JSONField, TextField
 from tortoise.models import Model
@@ -25,18 +16,8 @@ from tortoise.models import Model
 import faebryk.library._F as F
 from faebryk.core.module import Module
 from faebryk.core.parameter import (
-    Numbers,
     Parameter,
     ParameterOperatable,
-)
-from faebryk.core.solver.solver import LOG_PICK_SOLVE, Solver
-from faebryk.libs.e_series import E_SERIES
-from faebryk.libs.picker.common import (
-    PickerESeriesIntersectionError,
-    PickerUnboundedParameterError,
-    check_compatible_parameters,
-    generate_si_values,
-    try_attach,
 )
 from faebryk.libs.picker.lcsc import (
     LCSC_Part,
@@ -48,12 +29,10 @@ from faebryk.libs.picker.picker import (
 )
 from faebryk.libs.sets.quantity_sets import (
     Quantity_Interval,
-    Quantity_Interval_Disjoint,
     Quantity_Singleton,
 )
-from faebryk.libs.sets.sets import P_UnitSet
 from faebryk.libs.units import UndefinedUnitError, quantity
-from faebryk.libs.util import at_exit, once
+from faebryk.libs.util import once
 
 logger = logging.getLogger(__name__)
 
@@ -383,438 +362,3 @@ class Component(Model):
             return self.manufacturer_name
         except AttributeError:
             return asyncio.run(Manufacturers().get_from_id(self.manufacturer_id))
-
-
-class ComponentQuery:
-    class Error(Exception): ...
-
-    class ParamError(Error):
-        def __init__(self, param: P_UnitSet, msg: str):
-            self.param = param
-            self.msg = msg
-            super().__init__(f"{msg} for parameter {param!r}")
-
-    def __init__(self):
-        # init db connection
-        JLCPCB_DB()
-
-        self.Q: Q | None = Q()
-        self.results: list[Component] | None = None
-
-    async def exec(self) -> list[Component]:
-        assert self.Q
-        queryset = Component.filter(self.Q)
-        logger.debug(f"Query results: {await queryset.count()}")
-        self.results = await queryset
-        self.Q = None
-        return self.results
-
-    def get(self) -> list[Component]:
-        if self.results is not None:
-            return self.results
-        return asyncio.run(self.exec())
-
-    def filter_by_stock(self, qty: int) -> Self:
-        assert self.Q
-        self.Q &= Q(stock__gte=qty)
-        return self
-
-    def filter_by_description(self, *keywords: str) -> Self:
-        assert self.Q
-
-        logger.debug(f"Possible keywords: {keywords}")
-        description_query = Q()
-        for keyword in keywords:
-            description_query |= Q(description__contains=keyword)
-        self.Q &= description_query
-
-        return self
-
-    def filter_by_si_values(
-        self,
-        value: Quantity_Interval_Disjoint,
-        e_series: E_SERIES | None = None,
-        tolerance_requirement: float | None = None,
-    ) -> Self:
-        assert self.Q
-        assert not self.results
-
-        try:
-            si_vals = generate_si_values(value, e_series)
-        except PickerUnboundedParameterError:
-            return self
-        except PickerESeriesIntersectionError as e:
-            raise ComponentQuery.ParamError(value, str(e)) from e
-
-        if tolerance_requirement:
-            self.filter_by_tolerance(tolerance_requirement)
-
-        return self.filter_by_description(*si_vals)
-
-    def hint_filter_parameter(
-        self,
-        param: Parameter,
-        solver: Solver,
-        e_series: E_SERIES | None = None,
-    ) -> Self:
-        # param will in the general case consist of multiple ranges
-        # we have to pick some range or make a new one to pre_filter our candidates
-        # we can try making a new range with inspect_min and max to filter out
-        # everything we already know won't fit
-        # then we can check the cardinality of the remaining candidates to see if we
-        # need to pick a range contained in the param to filter
-
-        # TODO
-        if not isinstance(param.domain, Numbers):
-            raise NotImplementedError()
-
-        if LOG_PICK_SOLVE:
-            logger.info(f"SQL filter: Getting known supersets for {param}")
-        candidate_ranges = solver.inspect_get_known_supersets(param)
-        if not candidate_ranges.is_finite():
-            return self
-        if not isinstance(candidate_ranges, Quantity_Interval_Disjoint):
-            # TODO nothing we can really do in the sqlite version here
-            # might just save it for later filtering
-            return self
-
-        return self.filter_by_si_values(candidate_ranges, e_series)
-
-    def filter_by_tolerance(self, tolerance: float) -> Self:
-        assert self.Q
-
-        tol_int = int(math.floor(tolerance * 100))
-        return self.filter_by_description(f"±{tol_int}%")
-
-    def filter_by_category(self, category: str, subcategory: str) -> Self:
-        return self.filter_by_categories([(category, subcategory)])
-
-    def filter_by_categories(self, categories: list[tuple[str, str]]) -> Self:
-        assert self.Q
-        category_ids = [
-            c_id
-            for category, subcategory in categories
-            for c_id in asyncio.run(Category().get_ids(category, subcategory))
-        ]
-        self.Q &= Q(category_id__in=category_ids)
-        return self
-
-    def filter_by_package(self, package_candidates: Sequence[str] | None) -> Self:
-        assert self.Q
-        if not package_candidates:
-            return self
-        package_query = Q()
-        if package_candidates is not None:
-            for package in package_candidates:
-                package_query |= Q(description__icontains=package)
-        self.Q &= package_query
-        return self
-
-    def filter_by_traits(self, obj: Module) -> Self:
-        out = self
-        if obj.has_trait(F.has_package_requirement):
-            out = self.filter_by_package(
-                obj.get_trait(F.has_package_requirement).get_package_candidates()
-            )
-
-        return out
-
-    def sort_by_price(self, qty: int = 1) -> Self:
-        self.get().sort(key=lambda x: x.get_price(qty))
-        return self
-
-    def filter_by_lcsc_pn(self, partnumber: str) -> Self:
-        assert self.Q
-        self.Q &= Q(lcsc=partnumber.strip("C"))
-        return self
-
-    def filter_by_specified_parameters(self, mapping: list[MappingParameterDB]) -> Self:
-        assert self.Q
-        keys = [m.attr_keys for m in mapping]
-
-        extra_query = Q()
-        for kl in keys:
-            sub_q = Q()
-            for k in kl:
-                sub_q |= Q(extra__contains=k)
-            extra_query &= sub_q
-        self.Q &= extra_query
-        return self
-
-    def filter_by_attribute_mention(self, candidates: list[str]) -> Self:
-        if not candidates:
-            return self
-        assert self.Q
-        q = Q()
-        for candidate in candidates:
-            q |= Q(extra__contains=candidate)
-        self.Q &= q
-        return self
-
-    def filter_by_manufacturer_pn(self, partnumber: str) -> Self:
-        assert self.Q
-        self.Q &= Q(mfr__icontains=partnumber)
-        return self
-
-    def filter_by_manufacturer(self, manufacturer: str) -> Self:
-        assert self.Q
-        if not manufacturer:
-            return self
-        manufacturer_ids = asyncio.run(Manufacturers().get_ids(manufacturer))
-        self.Q &= Q(manufacturer_id__in=manufacturer_ids)
-        return self
-
-    def filter_by_module_params(
-        self,
-        module: Module,
-        mapping: list[MappingParameterDB],
-        solver: Solver,
-    ) -> Generator[Component, None, None]:
-        """
-        Filter the results by the parameters of the module
-
-        This should be used as the last step before attaching the component to the
-        module
-
-        :param module: The module to filter by
-        :param mapping: The mapping of module parameters to component attributes
-        :param qty: The quantity of components needed
-        :param attach_first: Whether to attach the first component that matches the
-        parameters and return immediately
-
-        :return: The first component that matches the parameters
-        """
-
-        # iterate through all candidate components
-        for c in self.get():
-            if check_compatible_parameters(module, c, mapping, solver):
-                yield c
-
-    def filter_by_module_params_and_attach(
-        self,
-        module: Module,
-        mapping: list[MappingParameterDB],
-        solver: Solver,
-        qty: int = 1,
-    ):
-        try_attach(
-            module,
-            self.filter_by_module_params(module, mapping, solver),
-            mapping,
-            qty,
-        )
-
-
-class JLCPCB_DB:
-    @dataclass
-    class Config:
-        db_path: Path = CACHE_FOLDER / Path("jlcpcb_part_database")
-        no_download_prompt: bool = False
-        force_db_update: bool = False
-
-    config = Config()
-    _instance: "JLCPCB_DB | None" = None
-    failed: Exception | None = None
-
-    @staticmethod
-    def get() -> "JLCPCB_DB":
-        return JLCPCB_DB.__new__(JLCPCB_DB)
-
-    def __new__(cls) -> "JLCPCB_DB":
-        if cls.failed:
-            raise cls.failed
-        if not JLCPCB_DB._instance:
-            instance = super(JLCPCB_DB, cls).__new__(cls)
-            try:
-                instance.init()
-            except FileNotFoundError as e:
-                cls.failed = e
-                raise e
-
-            JLCPCB_DB._instance = instance
-            at_exit(JLCPCB_DB.close)
-        return JLCPCB_DB._instance
-
-    @staticmethod
-    def close():
-        if not JLCPCB_DB._instance:
-            return
-        instance = JLCPCB_DB._instance
-        JLCPCB_DB._instance = None
-        del instance
-
-    def init(self) -> None:
-        config = self.config
-        self.db_path = config.db_path
-        self.db_file = config.db_path / Path("cache.sqlite3")
-        self.connected = False
-        self.fresh_db = False
-
-        no_download_prompt = config.no_download_prompt
-
-        if not sys.stdin.isatty():
-            no_download_prompt = True
-
-        if config.force_db_update:
-            self.download()
-        elif not self.has_db():
-            if no_download_prompt or self.prompt_db_update(
-                f"No JLCPCB database found at {self.db_file}, download now?"
-            ):
-                self.download()
-            else:
-                raise FileNotFoundError(f"No JLCPCB database found at {self.db_file}")
-        elif not self.is_db_up_to_date():
-            if not no_download_prompt and self.prompt_db_update(
-                f"JLCPCB database at {self.db_file} is older than 7 days, update?"
-            ):
-                self.download()
-            else:
-                logger.warning("Continuing with outdated JLCPCB database")
-
-        asyncio.run(self._init_db())
-
-    def __del__(self):
-        if self.connected:
-            asyncio.run(self._close_db())
-
-    async def _init_db(self):
-        await Tortoise.init(
-            db_url=f"sqlite://{self.db_path}/cache.sqlite3",
-            modules={
-                "models": [__name__]
-            },  # Use __name__ to refer to the current module
-        )
-        self.connected = True
-        if self.fresh_db:
-            await self.post_process_db()
-
-    async def _close_db(self):
-        from tortoise.log import logger as tortoise_logger
-
-        # suppress close ORM info
-        tortoise_logger.setLevel(logging.WARNING)
-        await Tortoise.close_connections()
-        self.connected = False
-
-    def has_db(self) -> bool:
-        return self.db_path.is_dir() and self.db_file.is_file()
-
-    def is_db_up_to_date(
-        self, max_timediff: datetime.timedelta = datetime.timedelta(days=7)
-    ) -> bool:
-        if not self.has_db():
-            return False
-
-        return (
-            datetime.datetime.fromtimestamp(
-                self.db_file.stat().st_mtime, tz=datetime.timezone.utc
-            )
-            >= datetime.datetime.now(tz=datetime.timezone.utc) - max_timediff
-        )
-
-    def prompt_db_update(self, prompt: str = "Update JLCPCB database?") -> bool:
-        ans = input(prompt + " [y/N]:").lower()
-        return ans == "y"
-
-    async def _remove_out_of_stock(self, non_passives: bool = False):
-        if not non_passives:
-            categories = [
-                ("Resistors", ""),
-                ("Capacitors", ""),
-                ("Inductors", ""),
-                ("Diodes", ""),
-            ]
-            category_ids = [
-                ids
-                for category, subcategory in categories
-                for ids in await Category().get_ids(category, subcategory)
-            ]
-            logger.info("Deleting out-of-stock passives from DB")
-            await Component.filter(category_id__in=category_ids, stock__lt=1).delete()
-
-        else:
-            logger.info("Deleting out-of-stock components from DB")
-            await Component.filter(stock__lt=1).delete()
-
-    async def post_process_db(self):
-        # Ignoring all OOS components isn't a good idea, since there are many
-        # parts you may want to explicitly include in your BoM, even when OOS
-        # eg. most ICs, micros
-        # TODO: consider another approach to optimize the DB, eg. partitioning
-        await self._remove_out_of_stock(non_passives=False)
-
-        logger.info("Vacuuming DB")
-        await Tortoise.get_connection("default").execute_query("VACUUM;")
-
-    def download(
-        self,
-    ):
-        def download_file(url, output_path: Path):
-            with requests.get(url, stream=True) as r:
-                r.raise_for_status()
-                with open(output_path, "wb") as f:
-                    for chunk in r.iter_content(chunk_size=8192):
-                        f.write(chunk)
-
-        def get_number_of_volumes(zip_path):
-            with open(zip_path, "rb") as f:
-                f.seek(-22, os.SEEK_END)  # Go to the end of the file minus 22 bytes
-                end_of_central_dir = f.read(22)
-
-                if len(end_of_central_dir) != 22 or not end_of_central_dir.startswith(
-                    b"PK\x05\x06"
-                ):
-                    # Not a valid ZIP file or the end of central directory signature is
-                    # missing
-                    raise ValueError(
-                        "Invalid ZIP file or End of Central Directory signature not "
-                        "found"
-                    )
-
-                # Unpack the number of this volume (should be 0 if single part zip)
-                current_volume, volume_with_central_dir = struct.unpack(
-                    "<HH", end_of_central_dir[4:8]
-                )
-
-                # Number of volume files = volume_with_central_dir + 1
-                return volume_with_central_dir + 1
-
-        self.db_path.mkdir(parents=True, exist_ok=True)
-
-        zip_file = self.db_path / Path("cache.zip")
-        base_url = "https://yaqwsx.github.io/jlcparts/data/"
-
-        logger.info(f"Downloading {base_url}cache.zip to {zip_file}")
-        download_file(base_url + "cache.zip", zip_file)
-
-        num_volumes = get_number_of_volumes(zip_file)
-        assert num_volumes <= 99
-        logger.info(f"Number of volumes: {num_volumes}")
-
-        # Download the additional volume files
-        for volume_num in track(
-            range(num_volumes), description="Downloading and appending zip volumes"
-        ):
-            # Skip .zip file since it is already downloaded
-            if volume_num == 0:
-                continue
-
-            volume_file = self.db_path / Path(f"cache.z{volume_num:02d}")
-            volume_url = base_url + f"cache.z{volume_num:02d}"
-
-            download_file(volume_url, volume_file)
-
-        self.db_file.unlink(missing_ok=True)
-        logger.info(f"Unzipping {zip_file}")
-        patoolib.extract_archive(str(zip_file), outdir=str(self.db_path))
-
-        # remove downloaded files
-        for volume_num in range(num_volumes):
-            if volume_num == 0:
-                volume_file = zip_file
-            else:
-                volume_file = self.db_path / Path(f"cache.z{volume_num:02d}")
-            os.remove(volume_file)
-
-        self.fresh_db = True
