+++ test/core/solver/test_solver.py
@@ -28,7 +28,6 @@ from faebryk.libs.app.parameters import resolve_dynamic_parameters
 from faebryk.libs.library import L
 from faebryk.libs.library.L import Range, RangeWithGaps, Single
 from faebryk.libs.picker.api.pickers import add_api_pickers
-from faebryk.libs.picker.jlcpcb.pickers import add_jlcpcb_pickers
 from faebryk.libs.picker.lcsc import LCSC_Part
 from faebryk.libs.picker.picker import (
     PickerOption,
@@ -681,7 +680,7 @@ def test_jlcpcb_pick_powered_led():
     children_mods = led.get_children_modules(direct_only=False, types=(Module,))
     for mod in children_mods:
         # add_api_pickers(mod)
-        add_jlcpcb_pickers(mod)
+        add_api_pickers(mod)
 
     pick_part_recursively(led, solver)
